+++ qrenderdoc/Windows/PipelineState/VulkanPipelineStateViewer.cpp
@@ -1965,22 +1965,22 @@ void VulkanPipelineStateViewer::setState()
   ui->stencils->clear();
   if(state.depthStencil.stencilTestEnable)
   {
-    ui->stencils->addTopLevelItem(
-        new RDTreeWidgetItem({tr("Front"), ToQStr(state.depthStencil.frontFace.function),
-                              ToQStr(state.depthStencil.frontFace.failOperation),
-                              ToQStr(state.depthStencil.frontFace.depthFailOperation),
-                              ToQStr(state.depthStencil.frontFace.passOperation),
-                              Formatter::Format(state.depthStencil.frontFace.writeMask, true),
-                              Formatter::Format(state.depthStencil.frontFace.compareMask, true),
-                              Formatter::Format(state.depthStencil.frontFace.reference, true)}));
-    ui->stencils->addTopLevelItem(
-        new RDTreeWidgetItem({tr("Back"), ToQStr(state.depthStencil.backFace.function),
-                              ToQStr(state.depthStencil.backFace.failOperation),
-                              ToQStr(state.depthStencil.backFace.depthFailOperation),
-                              ToQStr(state.depthStencil.backFace.passOperation),
-                              Formatter::Format(state.depthStencil.backFace.writeMask, true),
-                              Formatter::Format(state.depthStencil.backFace.compareMask, true),
-                              Formatter::Format(state.depthStencil.backFace.reference, true)}));
+    ui->stencils->addTopLevelItem(new RDTreeWidgetItem(
+        {tr("Front"), ToQStr(state.depthStencil.frontFace.function),
+         ToQStr(state.depthStencil.frontFace.failOperation),
+         ToQStr(state.depthStencil.frontFace.depthFailOperation),
+         ToQStr(state.depthStencil.frontFace.passOperation),
+         Formatter::Format((uint8_t)state.depthStencil.frontFace.writeMask, true),
+         Formatter::Format((uint8_t)state.depthStencil.frontFace.compareMask, true),
+         Formatter::Format((uint8_t)state.depthStencil.frontFace.reference, true)}));
+    ui->stencils->addTopLevelItem(new RDTreeWidgetItem(
+        {tr("Back"), ToQStr(state.depthStencil.backFace.function),
+         ToQStr(state.depthStencil.backFace.failOperation),
+         ToQStr(state.depthStencil.backFace.depthFailOperation),
+         ToQStr(state.depthStencil.backFace.passOperation),
+         Formatter::Format((uint8_t)state.depthStencil.backFace.writeMask, true),
+         Formatter::Format((uint8_t)state.depthStencil.backFace.compareMask, true),
+         Formatter::Format((uint8_t)state.depthStencil.backFace.reference, true)}));
   }
   else
   {
